I included all of the code that I had made up until this point.
All of my work in this push has been for the Factory level.

>Powerups pickup
>Powerups make sounds when picked up
>Players who fall into the lava will die and be respawned at the checkpoint.
>Falling into the lava plays a small sound.

Issues: The only issue so far is that the player will fall into the lava, explode, and then be teleported to the center of the map to once again fall into the lava. After this second time the player is transported to the intended respawn location.